Change parser behavior; refactoring
To allow infix parsers to have the same syntax (shape) as
prefix parsers, it was necessary to update ParseContext to
always have the current node. This drove a change in the
parser to use the ParseContext to get the next prefix/infix
node so it could update current (avoiding the mutator).

All this so that RequireNode would have visibility to the
current node.

Create algebraic transformations to turn infix parsers into
lambdas that use NodeParseRule's.